fix(eslint,vitest,auth,me,map): address PR #873 review comments
import/no-cycle was a silent no-op (import-x's default resolver never
traverses .ts/.tsx), and vitest/expect-expect was disabled outright
rather than tuned. Add the missing resolver settings and fix the one
real cycle it surfaces (email-provider.ts <-> utils.ts) by relocating
sendVerificationRequest alongside the provider config that consumes
it. Enable expect-expect with assertFunctionNames tuning and fix the
one genuinely empty test it catches, with targeted disables on tests
that intentionally assert no state update occurs.
