+++ apps/map/__tests__/app/user.test.tsx
@@ -3,8 +3,7 @@ import "@testing-library/jest-dom";
 import { describe, it } from "vitest";
 
 describe("UserComponent", () => {
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  it("should render the user component", () => {});
+  it.todo("should render the user component");
 
   // Add more tests here
 });
